cosmetics: rename the various *shoppingCart* things
That the names don't clash so often.
