+++ app/src/components/Navigation/ShoppingCartIcon.tsx
@@ -1,5 +1,5 @@
 // noinspection ES6PreferShortImport
-import { numberOfProducts } from '../../store/shoppingCart';
+import { numberOfProducts } from '../../store/shoppingCartStore';
 import { Avatar, AvatarBadge, Text } from '@chakra-ui/react';
 import { CartShopping } from '@styled-icons/fa-solid';
 import { useAtom } from 'jotai/react';
@@ -8,7 +8,7 @@ import React, { CSSProperties } from 'react';
 type ShoppingCartProperties = {
   style?: CSSProperties;
 };
-export function ShoppingCart({ style = {} }: ShoppingCartProperties) {
+export function ShoppingCartIcon({ style = {} }: ShoppingCartProperties) {
   const [numberOfProductsValue] = useAtom(numberOfProducts);
   const showMaxOf9 = numberOfProductsValue >= 9 ? '9+' : numberOfProductsValue;
   return (
